Added functionality to add player and view players

+++ database_sql_scripts/Database_Clear.sql
@@ -1,10 +1,16 @@
--- Clear all tables --
-DROP TABLE Plays_For;
+-- Clear all tables if they exist --
+IF OBJECT_ID('Plays_For', 'U') IS NOT NULL
+    DROP TABLE Plays_For;
 
-DROP TABLE Power_Users;
+IF OBJECT_ID('Power_Users', 'U') IS NOT NULL
+    DROP TABLE Power_Users;
 
-DROP TABLE Teams;
+IF OBJECT_ID('Teams', 'U') IS NOT NULL
+    DROP TABLE Teams;
 
-DROP TABLE Players;
+IF OBJECT_ID('Players', 'U') IS NOT NULL
+    DROP TABLE Players;
 
-DROP TABLE Countries;
+IF OBJECT_ID('Countries', 'U') IS NOT NULL
+    DROP TABLE Countries;
+
